+++ apps/bfd-server/bfd-server-shared-utils/pom.xml
@@ -1,0 +1,28 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+
+	<parent>
+		<groupId>gov.cms.bfd</groupId>
+		<artifactId>bfd-server-parent</artifactId>
+		<version>1.0.0-SNAPSHOT</version>
+	</parent>
+
+	<artifactId>bfd-server-shared-utils</artifactId>
+	<packaging>jar</packaging>
+
+	<description>
+		A Java JAR and assembly zip that packages up shared utilities used by
+		the FHIR Server and Launcher.
+	</description>
+
+	<dependencies>
+		<dependency>
+			<!-- Used to run our unit and integration tests. -->
+			<groupId>org.junit.jupiter</groupId>
+			<artifactId>junit-jupiter-engine</artifactId>
+			<scope>test</scope>
+		</dependency>
+	</dependencies>
+
+</project>
