+++ apps/bfd-server/bfd-server-shared-utils/src/main/java/gov/cms/bfd/server/sharedutils/MDCFormatter.java
@@ -1,4 +1,4 @@
-package gov.cms.bfd.sharedutils.logging;
+package gov.cms.bfd.server.sharedutils;
 
 /** Formatter for logging to the MDC class. */
 public class MDCFormatter {
