+++ apps/bfd-server/bfd-server-shared-utils/src/test/java/gov/cms/bfd/server/sharedutils/MDCFormatterTest.java
@@ -1,4 +1,4 @@
-package gov.cms.bfd.sharedutils.logging;
+package gov.cms.bfd.server.sharedutils;
 
 import static org.junit.jupiter.api.Assertions.*;
