+++ apps/bfd-server/bfd-server-war/pom.xml
@@ -45,7 +45,7 @@
 		<dependency>
 			<!-- Includes a formatter around MDC for logging purposes -->
 			<groupId>gov.cms.bfd</groupId>
-			<artifactId>bfd-shared-utils</artifactId>
+			<artifactId>bfd-server-shared-utils</artifactId>
 			<version>1.0.0-SNAPSHOT</version>
 		</dependency>
 		<dependency>
