+++ apps/bfd-server/pom.xml
@@ -23,6 +23,7 @@
 	<modules>
 		<module>bfd-server-launcher-sample</module>
 		<module>bfd-server-launcher</module>
+		<module>bfd-server-shared-utils</module>
 		<module>bfd-server-war</module>
 	</modules>
 
@@ -65,16 +66,6 @@
 			</dependency>
 		</dependencies>
 	</dependencyManagement>
-
-	<dependencies>
-		<dependency>
-			<!-- Provides utility code and dependencies that are shared by (more or
-				less) all of the other BFD modules. -->
-			<groupId>gov.cms.bfd</groupId>
-			<artifactId>bfd-shared-utils</artifactId>
-			<version>${project.version}</version>
-		</dependency>
-	</dependencies>
 
 	<build>
 		<pluginManagement>
